--- src/main/java/executor/service/service/ExecutionService.java
@@ -1,10 +1,0 @@
-package executor.service.service;
-
-import executor.service.model.ProxyConfigHolder;
-import executor.service.model.Scenario;
-
-public interface ExecutionService {
-
-    void execute(Scenario scenario, ProxyConfigHolder proxy);
-
-}
