--- src/main/java/executor/service/service/ItemHandler.java
@@ -1,7 +1,0 @@
-package executor.service.service;
-
-public interface ItemHandler<T> {
-
-    void onItemReceived(T item);
-
-}
