--- src/main/java/executor/service/service/ProxySourcesClient.java
@@ -1,7 +1,0 @@
-package executor.service.service;
-
-public interface ProxySourcesClient {
-
-    void execute(ItemHandler handler);
-
-}
