--- src/main/java/executor/service/service/ProxyValidator.java
@@ -1,8 +1,0 @@
-package executor.service.service;
-
-import executor.service.model.ProxyConfigHolder;
-
-public interface ProxyValidator {
-
-    Boolean isValid(ProxyConfigHolder proxyConfigHolder);
-}
