--- src/main/java/executor/service/service/ScenarioExecutor.java
@@ -1,10 +1,0 @@
-package executor.service.service;
-
-import executor.service.model.Scenario;
-import org.openqa.selenium.WebDriver;
-
-public interface ScenarioExecutor {
-
-    void execute(Scenario scenario, WebDriver webDriver);
-
-}
