--- src/main/java/executor/service/service/ScenarioProvider.java
@@ -1,10 +1,0 @@
-package executor.service.service;
-
-import executor.service.model.Scenario;
-
-import java.util.List;
-
-public interface ScenarioProvider {
-
-    List<Scenario> readScenarios();
-}
