--- src/main/java/executor/service/service/ScenarioSourceListener.java
@@ -1,7 +1,0 @@
-package executor.service.service;
-
-public interface ScenarioSourceListener {
-
-    void execute(ItemHandler handler);
-
-}
