--- src/main/java/executor/service/service/impl/StepExecutionClickCssImpl.java
@@ -1,22 +1,0 @@
-package executor.service.service.impl;
-
-import executor.service.model.Step;
-import executor.service.model.StepAction;
-import executor.service.service.StepExecutionClickCss;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-
-public class StepExecutionClickCssImpl implements StepExecutionClickCss {
-
-  @Override
-  public String getStepAction() {
-    return StepAction.CLICK_CSS.getName();
-  }
-
-  @Override
-  public void step(WebDriver webDriver, Step step) {
-    WebElement element = webDriver.findElement(By.cssSelector(step.getValue()));
-    element.click();
-  }
-}
