--- src/main/java/executor/service/service/impl/StepExecutionSleepImpl.java
@@ -1,41 +1,0 @@
-package executor.service.service.impl;
-
-import executor.service.exceptions.StepExecutionException;
-import executor.service.model.Step;
-import executor.service.model.StepAction;
-import executor.service.service.StepExecutionSleep;
-import org.openqa.selenium.WebDriver;
-
-import java.util.Random;
-
-public class StepExecutionSleepImpl implements StepExecutionSleep {
-
-  @Override
-  public String getStepAction() {
-    return StepAction.SLEEP.getName();
-  }
-
-  @Override
-  public void step(WebDriver webDriver, Step step) {
-    try {
-      Thread.sleep(getRandom(step.getValue()));
-    } catch (InterruptedException e) {
-      throw new StepExecutionException(e);
-    }
-
-  }
-
-  private int getRandom(java.lang.String value) {
-
-    java.lang.String values[] = value.split(":");
-    Random random = new Random();
-    int min = Integer.parseInt(values[0]);
-    int max = Integer.parseInt(values[values.length - 1]);
-
-    return random.ints(min, max + 1)
-        .findFirst().orElse(1);
-
-  }
-
-
-}
