--- src/main/resources/proxy-credential.json
@@ -1,18 +1,0 @@
-[
-  {
-    "username": "user11",
-    "password": "pass1"
-  },
-  {
-    "username": "user2",
-    "password": "pass2"
-  },
-  {
-    "username": "user3",
-    "password": "pass13"
-  },
-  {
-    "username": "user4",
-    "password": "pass4"
-  }
-]
