--- src/main/resources/proxy-credentials-default.json
@@ -1,6 +1,0 @@
-[
-  {
-    "username": "user2",
-    "password": "pass2"
-  }
-]
