--- src/main/resources/proxy-network-default.json
@@ -1,6 +1,0 @@
-[
-  {
-    "hostname": "host1",
-    "port": 8080
-  }
-]
