--- src/main/resources/proxy-network.json
@@ -1,14 +1,0 @@
-[
-  {
-    "hostname": "host1",
-    "port": 8080
-  },
-  {
-    "hostname": "host2",
-    "port": 8088
-  },
-  {
-    "hostname": "host3",
-    "port": 8089
-  }
-]
