--- src/main/resources/proxy-validator.properties
@@ -1,2 +1,0 @@
-target.url=https://www.google.com
-connection.timeout=10000
