--- src/main/resources/thread-pool.properties
@@ -1,2 +1,0 @@
-core.pool.size=2
-keep.alive.time=10
